ePaperless_task_schedule/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
argcomplete==1.10.3
asgiref==3.5.0
asyncio==3.4.3
beautifulsoup4==4.8.2
blinker==1.4
boto3==1.34.90
botocore==1.34.90
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
command-not-found==0.3
compressed-rtf==1.0.6
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
docx2txt==0.8
ebcdic==1.1.1
exceptiongroup==1.2.1
extract-msg==0.28.7
fastapi==0.110.2
h11==0.14.0
httplib2==0.20.2
httptools==0.6.1
idna==3.7
IMAPClient==2.1.0
importlib-metadata==4.6.4
jeepney==0.7.1
jmespath==1.0.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==5.2.1
more-itertools==8.10.0
netifaces==0.11.0
numpy==1.21.5
oauthlib==3.2.0
olefile==0.47
packaging==24.0
pdf2image==1.17.0
pdfminer.six==20191110
pillow==10.3.0
pycairo==1.26.0
pycryptodome==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
PyGObject==3.42.1
PyJWT==2.3.0
PyMuPDF==1.24.2
PyMuPDFb==1.24.1
pyparsing==2.4.7
pytesseract==0.3.10
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-pptx==0.6.23
PyYAML==5.4.1
requests==2.31.0
s3transfer==0.10.1
SecretStorage==3.3.1
six==1.12.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.8.1
starlette==0.37.2
systemd-python==234
textract==1.6.5
typing_extensions==4.11.0
tzlocal==5.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
wadllib==1.3.6
watchfiles==0.21.0
websockets==12.0
wsproto==1.0.0
xlrd==1.2.0
XlsxWriter==3.2.0
zipp==1.0.0
